+++ packages/vtable/examples/menu.ts
@@ -48,6 +48,10 @@ export const menus = [
       },
       {
         path: 'list',
+        name: 'list-more-columns'
+      },
+      {
+        path: 'list',
         name: 'list-no-columns'
       }
     ]
@@ -147,6 +151,10 @@ export const menus = [
       {
         path: 'pivot-chart',
         name: 'pivot-chart-event'
+      },
+      {
+        path: 'pivot-chart',
+        name: 'pivotChart-aeolus-large-column'
       }
     ]
   },