+++ packages/vtable/src/scenegraph/layout/update-row.ts
@@ -1,6 +1,6 @@
 import type { CellAddress } from '../../ts-types';
 import type { BaseTableAPI } from '../../ts-types/base-table';
-import type { Group } from '../graphic/group';
+import { Group } from '../graphic/group';
 import { updateCell } from '../group-creater/cell-helper';
 import type { Scenegraph } from '../scenegraph';
 
@@ -40,14 +40,26 @@ export function updateRow(
   });
 
   // reset attribute y and row number in CellGroup
-  resetRowNumberAndY(scene);
+  const newTotalHeight = resetRowNumberAndY(scene);
+
+  if (addRows.length) {
+    scene.proxy.rowUpdatePos = scene.proxy.rowStart;
+    scene.proxy.rowUpdateDirection = 'up';
+    scene.proxy.updateCellGroups(scene.proxy.screenRowCount * 2);
+    scene.proxy.progress();
+  }
+
+  // update table size
+  scene.updateContainerHeight(scene.table.frozenRowCount, newTotalHeight - scene.bodyGroup.attribute.height);
 }
 
 function removeRow(row: number, scene: Scenegraph) {
   for (let col = 0; col < scene.table.colCount; col++) {
     // const headerColGroup = scene.getColGroup(col, true);
     const colGroup = scene.getColGroup(col, false);
-
+    if (!colGroup) {
+      continue;
+    }
     // // remove cellGroup in headerColGroup
     // let headerCellGroup;
     // headerColGroup.forEachChildren((cellGroup: Group) => {
@@ -77,13 +89,24 @@ function removeRow(row: number, scene: Scenegraph) {
   // TODO 需要整体更新proxy的状态
   scene.proxy.rowEnd--;
   scene.proxy.bodyBottomRow--;
+  scene.proxy.currentRow--;
   scene.proxy.totalRow--;
 }
 
 function addRow(row: number, scene: Scenegraph) {
   for (let col = 0; col < scene.table.colCount; col++) {
     // create cellGroup
-    const cellGroup = updateCell(col, row, scene.table, true);
+    // const cellGroup = updateCell(col, row, scene.table, true);
+    const cellGroup = new Group({
+      x: 0,
+      y: 0,
+      width: scene.table.getColWidth(col),
+      height: scene.table.getRowHeight(row)
+    });
+    cellGroup.role = 'cell';
+    cellGroup.col = col;
+    cellGroup.row = row;
+    cellGroup.needUpdate = true;
 
     if (!cellGroup) {
       // TODO: deal with data promise situation
@@ -92,34 +115,46 @@ function addRow(row: number, scene: Scenegraph) {
 
     // insert cellGroup in colGroup
     const colGroup = scene.getColGroup(col);
+    if (!colGroup) {
+      continue;
+    }
+
     if (colGroup.firstChild && row < (colGroup.firstChild as Group).row) {
       colGroup.insertBefore(cellGroup, colGroup.firstChild);
+      (colGroup.firstChild as Group).row = (colGroup.firstChild as Group).row + 1;
     } else if (colGroup.lastChild && row > (colGroup.lastChild as Group).row) {
       colGroup.appendChild(cellGroup);
     } else {
-      let cellBefore;
-      colGroup.forEachChildren((cellGroup: Group) => {
-        if (cellGroup.row === row) {
-          cellBefore = cellGroup;
-          return true;
-        }
-        return false;
-      });
+      // let cellBefore: Group;
+      // colGroup.forEachChildren((cellGroup: Group) => {
+      //   if (cellGroup.row === row) {
+      //     cellBefore = cellGroup;
+      //     return true;
+      //   }
+      //   return false;
+      // });
+
+      const cellBefore = scene.highPerformanceGetCell(col, row, true);
       if (cellBefore !== cellGroup) {
         colGroup.insertBefore(cellGroup, cellBefore);
+        cellBefore && (cellBefore.row = cellBefore.row + 1);
+        if (cellBefore !== colGroup.lastChild) {
+          colGroup.lastChild && ((colGroup.lastChild as Group).row = (colGroup.lastChild as Group).row + 1);
+        }
       }
     }
 
-    // reset row number
-    let rowIndex = (colGroup.firstChild as Group)?.row;
-    colGroup.forEachChildren((cellGroup: Group) => {
-      cellGroup.row = rowIndex;
-      rowIndex++;
-    });
+    // // reset row number
+    // let rowIndex = (colGroup.firstChild as Group)?.row;
+    // colGroup.forEachChildren((cellGroup: Group) => {
+    //   cellGroup.row = rowIndex;
+    //   rowIndex++;
+    // });
   }
   // TODO 需要整体更新proxy的状态
   scene.proxy.rowEnd++;
   scene.proxy.bodyBottomRow++;
+  scene.proxy.currentRow++;
   scene.proxy.totalRow++;
 }
 
@@ -138,6 +173,9 @@ function resetRowNumber(scene: Scenegraph) {
   for (let col = 0; col < scene.table.colCount; col++) {
     const headerColGroup = scene.getColGroup(col, true);
     const colGroup = scene.getColGroup(col, false);
+    if (!headerColGroup || !colGroup) {
+      continue;
+    }
     // reset row number
     let rowIndex = (headerColGroup.firstChild as Group)?.row;
     headerColGroup.forEachChildren((cellGroup: Group) => {
@@ -157,6 +195,9 @@ function resetRowNumberAndY(scene: Scenegraph) {
   for (let col = 0; col < scene.table.colCount; col++) {
     const headerColGroup = scene.getColGroup(col, true);
     const colGroup = scene.getColGroup(col, false);
+    if (!headerColGroup || !colGroup) {
+      continue;
+    }
     // reset row number
     let rowIndex = (headerColGroup.firstChild as Group)?.row;
     let y = 0;
@@ -189,7 +230,5 @@ function resetRowNumberAndY(scene: Scenegraph) {
     //   scene.proxy.referenceRow = scene.proxy.rowStart + Math.floor((scene.proxy.rowEnd - scene.proxy.rowStart) / 2);
     // }
   }
-
-  // update table size
-  scene.updateContainerHeight(scene.table.frozenRowCount, newTotalHeight - scene.bodyGroup.attribute.height);
+  return newTotalHeight;
 }