+++ packages/vtable/src/scenegraph/select/update-select-border.ts
@@ -1,82 +1,160 @@
 import type { IRect } from '@visactor/vrender';
 import type { Scenegraph } from '../scenegraph';
-import type { CellLocation } from '../../ts-types';
+import type { CellSubLocation } from '../../ts-types';
 import { getCellMergeInfo } from '../utils/get-cell-merge';
 
 export function updateAllSelectComponent(scene: Scenegraph) {
-  scene.selectingRangeComponents.forEach((selectComp: { rect: IRect; role: CellLocation }, key: string) => {
+  scene.selectingRangeComponents.forEach((selectComp: { rect: IRect; role: CellSubLocation }, key: string) => {
     updateComponent(selectComp, key, scene);
   });
-  scene.selectedRangeComponents.forEach((selectComp: { rect: IRect; role: CellLocation }, key: string) => {
+  scene.selectedRangeComponents.forEach((selectComp: { rect: IRect; role: CellSubLocation }, key: string) => {
     updateComponent(selectComp, key, scene);
   });
 }
 
-function updateComponent(selectComp: { rect: IRect; role: CellLocation }, key: string, scene: Scenegraph) {
+function updateComponent(selectComp: { rect: IRect; role: CellSubLocation }, key: string, scene: Scenegraph) {
   const [startColStr, startRowStr, endColStr, endRowStr] = key.split('-');
-  const startCol = parseInt(startColStr, 10);
-  const startRow = parseInt(startRowStr, 10);
-  const endCol = parseInt(endColStr, 10);
-  const endRow = parseInt(endRowStr, 10);
-  let cellsBounds;
-  for (let i = startCol; i <= endCol; i++) {
-    for (let j = startRow; j <= endRow; j++) {
-      const cellGroup = scene.highPerformanceGetCell(i, j);
-      if (cellGroup.role !== 'cell') {
-        continue;
+  let startCol = parseInt(startColStr, 10);
+  let startRow = parseInt(startRowStr, 10);
+  let endCol = parseInt(endColStr, 10);
+  let endRow = parseInt(endRowStr, 10);
+  // const cellsBounds;
+  // 下面逻辑根据选中区域所属表格部分 来判断可视区域内容的选中单元格范围
+  let visibleCellRange;
+  switch (selectComp.role) {
+    case 'rowHeader':
+      visibleCellRange = scene.table.getBodyVisibleRowRange();
+      if (visibleCellRange) {
+        startRow = Math.max(startRow, visibleCellRange.rowStart);
+        endRow = Math.min(endRow, visibleCellRange.rowEnd);
       }
-      cellGroup.AABBBounds.width(); // hack: globalAABBBounds可能不会自动更新，这里强制更新一下
-      const bounds = cellGroup.globalAABBBounds;
-      if (!cellsBounds) {
-        cellsBounds = bounds;
-      } else {
-        cellsBounds.union(bounds);
+      break;
+    case 'columnHeader':
+      visibleCellRange = scene.table.getBodyVisibleCellRange();
+      if (visibleCellRange) {
+        startCol = Math.max(startCol, visibleCellRange.colStart);
+        endCol = Math.min(endCol, visibleCellRange.colEnd);
       }
-    }
+      break;
+    case 'cornerHeader':
+      break;
+    case 'bottomFrozen':
+      visibleCellRange = scene.table.getBodyVisibleCellRange();
+      if (visibleCellRange) {
+        startCol = Math.max(startCol, visibleCellRange.colStart);
+        endCol = Math.min(endCol, visibleCellRange.colEnd);
+      }
+      break;
+    case 'rightFrozen':
+      visibleCellRange = scene.table.getBodyVisibleCellRange();
+      if (visibleCellRange) {
+        startRow = Math.max(startRow, visibleCellRange.rowStart);
+        endRow = Math.min(endRow, visibleCellRange.rowEnd);
+      }
+      break;
+    case 'rightTopCorner':
+      break;
+    case 'leftBottomCorner':
+      break;
+    case 'rightBottomCorner':
+      break;
+    default:
+      visibleCellRange = scene.table.getBodyVisibleCellRange();
+      if (visibleCellRange) {
+        startRow = Math.max(startRow, visibleCellRange.rowStart);
+        endRow = Math.min(endRow, visibleCellRange.rowEnd);
+        startCol = Math.max(startCol, visibleCellRange.colStart);
+        endCol = Math.min(endCol, visibleCellRange.colEnd);
+      }
+      break;
   }
-  if (!cellsBounds) {
-    // 选中区域在实际单元格区域外，不显示选择框
-    selectComp.rect.setAttributes({
-      visible: false
-    });
-  } else {
-    selectComp.rect.setAttributes({
-      x: cellsBounds.x1 - scene.tableGroup.attribute.x,
-      y: cellsBounds.y1 - scene.tableGroup.attribute.y,
-      width: cellsBounds.width(),
-      height: cellsBounds.height(),
-      visible: true
-    });
-  }
+  const cellRange = scene.table.getCellRange(startCol, startRow);
+  const colsWidth = scene.table.getColsWidth(cellRange.start.col, endCol);
+  const rowsHeight = scene.table.getRowsHeight(cellRange.start.row, endRow);
+  const firstCellBound = scene.highPerformanceGetCell(startCol, startRow).globalAABBBounds;
+  // if (!cellsBounds) {
+  //   // 选中区域在实际单元格区域外，不显示选择框
+  //   selectComp.rect.setAttributes({
+  //     visible: false
+  //   });
+  // } else {
+  selectComp.rect.setAttributes({
+    x: firstCellBound.x1 - scene.tableGroup.attribute.x, //坐标xy在下面的逻辑中会做适当调整
+    y: firstCellBound.y1 - scene.tableGroup.attribute.y,
+    width: colsWidth,
+    height: rowsHeight,
+    visible: true
+  });
+  // }
 
   //#region 判断是不是按着表头部分的选中框 因为绘制层级的原因 线宽会被遮住一半，因此需要动态调整层级
   const isNearRowHeader =
     // scene.table.scrollLeft === 0 &&
     startCol === scene.table.frozenColCount;
+  const isNearRightRowHeader =
+    // scene.table.scrollLeft === 0 &&
+    endCol === scene.table.colCount - scene.table.rightFrozenColCount - 1;
   const isNearColHeader =
     // scene.table.scrollTop === 0 &&
     startRow === scene.table.frozenRowCount;
+  const isNearBottomColHeader =
+    // scene.table.scrollTop === 0 &&
+    endRow === scene.table.rowCount - scene.table.bottomFrozenRowCount - 1;
   if (
     (isNearRowHeader && selectComp.rect.attribute.stroke[3]) ||
-    (isNearColHeader && selectComp.rect.attribute.stroke[0])
+    (isNearRightRowHeader && selectComp.rect.attribute.stroke[1]) ||
+    (isNearColHeader && selectComp.rect.attribute.stroke[0]) ||
+    (isNearBottomColHeader && selectComp.rect.attribute.stroke[2])
   ) {
-    if (isNearRowHeader) {
+    if (isNearRowHeader && selectComp.rect.attribute.stroke[3]) {
       scene.tableGroup.insertAfter(
         selectComp.rect,
-        selectComp.role === 'columnHeader' ? scene.cornerHeaderGroup : scene.rowHeaderGroup
+        selectComp.role === 'columnHeader'
+          ? scene.cornerHeaderGroup
+          : selectComp.role === 'bottomFrozen'
+          ? scene.leftBottomCornerGroup
+          : scene.rowHeaderGroup
       );
     }
-    if (isNearColHeader) {
+
+    if (isNearBottomColHeader && selectComp.rect.attribute.stroke[2]) {
       scene.tableGroup.insertAfter(
         selectComp.rect,
-        selectComp.role === 'rowHeader' ? scene.cornerHeaderGroup : scene.colHeaderGroup
+        selectComp.role === 'rowHeader'
+          ? scene.leftBottomCornerGroup
+          : selectComp.role === 'rightFrozen'
+          ? scene.rightBottomCornerGroup
+          : scene.bottomFrozenGroup
       );
     }
+
+    if (isNearColHeader && selectComp.rect.attribute.stroke[0]) {
+      scene.tableGroup.insertAfter(
+        selectComp.rect,
+        selectComp.role === 'rowHeader'
+          ? scene.cornerHeaderGroup
+          : selectComp.role === 'rightFrozen'
+          ? scene.rightTopCornerGroup
+          : scene.colHeaderGroup
+      );
+    }
+    if (isNearRightRowHeader && selectComp.rect.attribute.stroke[1]) {
+      scene.tableGroup.insertAfter(
+        selectComp.rect,
+        selectComp.role === 'columnHeader'
+          ? scene.rightTopCornerGroup
+          : selectComp.role === 'bottomFrozen'
+          ? scene.rightBottomCornerGroup
+          : scene.rightFrozenGroup
+      );
+    }
+
     //#region 调整层级后 滚动情况下会出现绘制范围出界 如body的选中框 渲染在了rowheader上面，所有需要调整选中框rect的 边界
     if (
       selectComp.rect.attribute.x < scene.rowHeaderGroup.attribute.width &&
+      // selectComp.rect.attribute.x + selectComp.rect.attribute.width > scene.rowHeaderGroup.attribute.width &&
       scene.table.scrollLeft > 0 &&
-      (selectComp.role === 'body' || selectComp.role === 'columnHeader')
+      (selectComp.role === 'body' || selectComp.role === 'columnHeader' || selectComp.role === 'bottomFrozen')
     ) {
       selectComp.rect.setAttributes({
         x: selectComp.rect.attribute.x + (scene.rowHeaderGroup.attribute.width - selectComp.rect.attribute.x),
@@ -84,13 +162,36 @@ function updateComponent(selectComp: { rect: IRect; role: CellLocation }, key: s
       });
     }
     if (
+      // selectComp.rect.attribute.x < scene.rightFrozenGroup.attribute.x &&
+      scene.rightFrozenGroup.attribute.width > 0 && // right冻结列存在的情况下
+      scene.rightFrozenGroup.attribute.height > 0 &&
+      selectComp.rect.attribute.x + selectComp.rect.attribute.width > scene.rightFrozenGroup.attribute.x &&
+      (selectComp.role === 'body' || selectComp.role === 'columnHeader' || selectComp.role === 'bottomFrozen')
+    ) {
+      selectComp.rect.setAttributes({
+        x: selectComp.rect.attribute.x,
+        width: scene.rightFrozenGroup.attribute.x - selectComp.rect.attribute.x
+      });
+    }
+    if (
       selectComp.rect.attribute.y < scene.colHeaderGroup.attribute.height &&
       scene.table.scrollTop > 0 &&
-      (selectComp.role === 'body' || selectComp.role === 'rowHeader')
+      (selectComp.role === 'body' || selectComp.role === 'rowHeader' || selectComp.role === 'rightFrozen')
     ) {
       selectComp.rect.setAttributes({
         y: selectComp.rect.attribute.y + (scene.colHeaderGroup.attribute.height - selectComp.rect.attribute.y),
         height: selectComp.rect.attribute.height - (scene.colHeaderGroup.attribute.height - selectComp.rect.attribute.y)
+      });
+    }
+    if (
+      scene.bottomFrozenGroup.attribute.width > 0 &&
+      scene.bottomFrozenGroup.attribute.height > 0 &&
+      selectComp.rect.attribute.y + selectComp.rect.attribute.height > scene.bottomFrozenGroup.attribute.y &&
+      (selectComp.role === 'body' || selectComp.role === 'rowHeader' || selectComp.role === 'rightFrozen')
+    ) {
+      selectComp.rect.setAttributes({
+        y: selectComp.rect.attribute.y,
+        height: scene.bottomFrozenGroup.attribute.y - selectComp.rect.attribute.y
       });
     }
     //#endregion
@@ -103,7 +204,17 @@ function updateComponent(selectComp: { rect: IRect; role: CellLocation }, key: s
         ? scene.colHeaderGroup
         : selectComp.role === 'rowHeader'
         ? scene.rowHeaderGroup
-        : scene.cornerHeaderGroup
+        : selectComp.role === 'cornerHeader'
+        ? scene.cornerHeaderGroup
+        : selectComp.role === 'rightTopCorner'
+        ? scene.rightTopCornerGroup
+        : selectComp.role === 'rightFrozen'
+        ? scene.rightFrozenGroup
+        : selectComp.role === 'leftBottomCorner'
+        ? scene.leftBottomCornerGroup
+        : selectComp.role === 'bottomFrozen'
+        ? scene.bottomFrozenGroup
+        : scene.rightBottomCornerGroup
     );
   }
   //#endregion
@@ -183,25 +294,73 @@ export function updateCellSelectBorder(
   };
   extendSelectRange();
   //#endregion
-  scene.selectingRangeComponents.forEach((selectComp: { rect: IRect; role: CellLocation }, key: string) => {
+  scene.selectingRangeComponents.forEach((selectComp: { rect: IRect; role: CellSubLocation }, key: string) => {
     selectComp.rect.delete();
   });
   scene.selectingRangeComponents = new Map();
 
   let needRowHeader = false;
+  let needRightRowHeader = false; // 右侧冻结
   let needColumnHeader = false;
+  let needBottomColumnHeader = false; // 底部冻结
   let needBody = false;
   let needCornerHeader = false;
+  let needRightTopCornerHeader = false;
+  let needRightBottomCornerHeader = false;
+  let needLeftBottomCornerHeader = false;
   if (startCol <= scene.table.frozenColCount - 1 && startRow <= scene.table.frozenRowCount - 1) {
     needCornerHeader = true;
   }
-  if (startCol <= scene.table.frozenColCount - 1 && endRow >= scene.table.frozenRowCount) {
+  if (endCol >= scene.table.colCount - scene.table.rightFrozenColCount && startRow <= scene.table.frozenRowCount - 1) {
+    needRightTopCornerHeader = true;
+  }
+
+  if (startCol <= scene.table.frozenColCount - 1 && endRow >= scene.table.rowCount - scene.table.bottomFrozenRowCount) {
+    needLeftBottomCornerHeader = true;
+  }
+
+  if (
+    endCol >= scene.table.colCount - scene.table.rightFrozenColCount &&
+    endRow >= scene.table.rowCount - scene.table.bottomFrozenRowCount
+  ) {
+    needRightBottomCornerHeader = true;
+  }
+
+  if (
+    startCol <= scene.table.frozenColCount - 1 &&
+    endRow >= scene.table.frozenRowCount &&
+    startRow <= scene.table.rowCount - scene.table.bottomFrozenRowCount - 1
+  ) {
     needRowHeader = true;
   }
-  if (startRow <= scene.table.frozenRowCount - 1 && endCol >= scene.table.frozenColCount) {
+  if (
+    endCol >= scene.table.colCount - scene.table.rightFrozenColCount &&
+    endRow >= scene.table.frozenRowCount &&
+    startRow <= scene.table.rowCount - scene.table.bottomFrozenRowCount - 1
+  ) {
+    needRightRowHeader = true;
+  }
+
+  if (
+    startRow <= scene.table.frozenRowCount - 1 &&
+    endCol >= scene.table.frozenColCount &&
+    startCol <= scene.table.colCount - scene.table.rightFrozenColCount - 1
+  ) {
     needColumnHeader = true;
   }
-  if (endCol >= scene.table.frozenColCount && endRow >= scene.table.frozenRowCount) {
+  if (
+    endRow >= scene.table.rowCount - scene.table.bottomFrozenRowCount &&
+    endCol >= scene.table.frozenColCount &&
+    startCol <= scene.table.colCount - scene.table.rightFrozenColCount - 1
+  ) {
+    needBottomColumnHeader = true;
+  }
+  if (
+    startCol <= scene.table.colCount - scene.table.rightFrozenColCount - 1 &&
+    endCol >= scene.table.frozenColCount &&
+    startRow <= scene.table.rowCount - scene.table.bottomFrozenRowCount - 1 &&
+    endRow >= scene.table.frozenRowCount
+  ) {
     needBody = true;
   }
 
@@ -220,30 +379,105 @@ export function updateCellSelectBorder(
       strokeArray
     );
   }
+  if (needRightTopCornerHeader) {
+    const cornerStartCol = Math.max(startCol, scene.table.colCount - scene.table.rightFrozenColCount);
+    const cornerEndRow = Math.min(endRow, scene.table.frozenRowCount - 1);
+    const strokeArray = [true, true, !needRightRowHeader, !needColumnHeader];
+    scene.createCellSelectBorder(
+      cornerStartCol,
+      startRow,
+      endCol,
+      cornerEndRow,
+      'rightTopCorner',
+      `${startCol}${startRow}${endCol}${endRow}`,
+      strokeArray
+    );
+  }
+
+  if (needLeftBottomCornerHeader) {
+    const cornerEndCol = Math.min(endCol, scene.table.frozenColCount - 1);
+    const cornerStartRow = Math.max(startRow, scene.table.rowCount - scene.table.bottomFrozenRowCount);
+    const strokeArray = [!needRowHeader, !needBottomColumnHeader, true, true];
+    scene.createCellSelectBorder(
+      startCol,
+      cornerStartRow,
+      cornerEndCol,
+      endRow,
+      'leftBottomCorner',
+      `${startCol}${startRow}${endCol}${endRow}`,
+      strokeArray
+    );
+  }
+  if (needRightBottomCornerHeader) {
+    const cornerStartCol = Math.max(startCol, scene.table.colCount - scene.table.rightFrozenColCount);
+    const cornerStartRow = Math.max(startRow, scene.table.rowCount - scene.table.bottomFrozenRowCount);
+    const strokeArray = [!needRightRowHeader, true, true, !needBottomColumnHeader];
+    scene.createCellSelectBorder(
+      cornerStartCol,
+      cornerStartRow,
+      endCol,
+      endRow,
+      'rightBottomCorner',
+      `${startCol}${startRow}${endCol}${endRow}`,
+      strokeArray
+    );
+  }
   if (needColumnHeader) {
     const columnHeaderStartCol = Math.max(startCol, scene.table.frozenColCount);
+    const columnHeaderEndCol = Math.min(endCol, scene.table.colCount - scene.table.rightFrozenColCount - 1);
     const columnHeaderEndRow = Math.min(endRow, scene.table.frozenRowCount - 1);
-    const strokeArray = [true, true, !needBody, !needCornerHeader];
+    const strokeArray = [true, !needRightTopCornerHeader, !needBody, !needCornerHeader];
     scene.createCellSelectBorder(
       columnHeaderStartCol,
       startRow,
-      endCol,
+      columnHeaderEndCol,
       columnHeaderEndRow,
       'columnHeader',
       `${startCol}${startRow}${endCol}${endRow}`,
       strokeArray
     );
   }
+  if (needBottomColumnHeader) {
+    const columnHeaderStartCol = Math.max(startCol, scene.table.frozenColCount);
+    const columnHeaderEndCol = Math.min(endCol, scene.table.colCount - scene.table.rightFrozenColCount - 1);
+    const columnHeaderStartRow = Math.max(startRow, scene.table.rowCount - scene.table.bottomFrozenRowCount);
+    const strokeArray = [!needBody, !needRightBottomCornerHeader, true, !needLeftBottomCornerHeader];
+    scene.createCellSelectBorder(
+      columnHeaderStartCol,
+      columnHeaderStartRow,
+      columnHeaderEndCol,
+      endRow,
+      'bottomFrozen',
+      `${startCol}${startRow}${endCol}${endRow}`,
+      strokeArray
+    );
+  }
   if (needRowHeader) {
     const columnHeaderStartRow = Math.max(startRow, scene.table.frozenRowCount);
+    const columnHeaderEndRow = Math.min(endRow, scene.table.rowCount - scene.table.bottomFrozenRowCount - 1);
     const columnHeaderEndCol = Math.min(endCol, scene.table.frozenColCount - 1);
-    const strokeArray = [!needCornerHeader, !needBody, true, true];
+    const strokeArray = [!needCornerHeader, !needBody, !needLeftBottomCornerHeader, true];
     scene.createCellSelectBorder(
       startCol,
       columnHeaderStartRow,
       columnHeaderEndCol,
-      endRow,
+      columnHeaderEndRow,
       'rowHeader',
+      `${startCol}${startRow}${endCol}${endRow}`,
+      strokeArray
+    );
+  }
+  if (needRightRowHeader) {
+    const columnHeaderStartRow = Math.max(startRow, scene.table.frozenRowCount);
+    const columnHeaderEndRow = Math.min(endRow, scene.table.rowCount - scene.table.bottomFrozenRowCount - 1);
+    const columnHeaderStartCol = Math.max(startCol, scene.table.colCount - scene.table.rightFrozenColCount);
+    const strokeArray = [!needRightTopCornerHeader, true, !needRightBottomCornerHeader, !needBody];
+    scene.createCellSelectBorder(
+      columnHeaderStartCol,
+      columnHeaderStartRow,
+      endCol,
+      columnHeaderEndRow,
+      'rightFrozen',
       `${startCol}${startRow}${endCol}${endRow}`,
       strokeArray
     );
@@ -251,12 +485,14 @@ export function updateCellSelectBorder(
   if (needBody) {
     const columnHeaderStartCol = Math.max(startCol, scene.table.frozenColCount);
     const columnHeaderStartRow = Math.max(startRow, scene.table.frozenRowCount);
-    const strokeArray = [!needColumnHeader, true, true, !needRowHeader];
+    const columnHeaderEndCol = Math.min(endCol, scene.table.colCount - scene.table.rightFrozenColCount - 1);
+    const columnHeaderEndRow = Math.min(endRow, scene.table.rowCount - scene.table.bottomFrozenRowCount - 1);
+    const strokeArray = [!needColumnHeader, !needRightRowHeader, !needBottomColumnHeader, !needRowHeader];
     scene.createCellSelectBorder(
       columnHeaderStartCol,
       columnHeaderStartRow,
-      endCol,
-      endRow,
+      columnHeaderEndCol,
+      columnHeaderEndRow,
       'body',
       `${startCol}${startRow}${endCol}${endRow}`,
       strokeArray