+++ packages/vtable/src/scenegraph/utils/update-container.ts
@@ -1,7 +1,7 @@
 import type { Group } from '../graphic/group';
 
-export function updateContainerChildrenX(containerGroup: Group): number {
-  let x = 0;
+export function updateContainerChildrenX(containerGroup: Group, x: number): number {
+  // let x = 0;
   containerGroup.forEachChildrenSkipChild((column: Group, index) => {
     column.setAttribute('x', x);
     x += column.attribute.width;