+++ packages/vtable/src/ts-types/base-table.ts
@@ -187,6 +187,12 @@ export interface IBaseTableProtected {
   renderChartAsync?: boolean;
   // // 开启图表异步渲染 每批次渐进渲染图表个数
   // renderChartAsyncBatchCount?: number;
+
+  /**
+   * 'auto':和浏览器滚动行为一致 表格滚动到顶部/底部时 触发浏览器默认行为;
+   *  设置为 'none' 时, 表格滚动到顶部/底部时, 不再触发父容器滚动
+   * */
+  overscrollBehavior?: 'auto' | 'none';
 }
 export interface BaseTableConstructorOptions {
   // /** 指定表格的行数 */
@@ -322,9 +328,10 @@ export interface BaseTableConstructorOptions {
 
   legends?: ITableLegendOption;
   title?: ITitle;
-  //是否开启图表异步渲染
+
+  /** 是否开启图表异步渲染 */
   renderChartAsync?: boolean;
-  // 开启图表异步渲染 每批次渐进渲染图表个数
+  /** 开启图表异步渲染 每批次渐进渲染图表个数  默认是5个 */
   renderChartAsyncBatchCount?: number;
 
   // for nodejs
@@ -332,6 +339,11 @@ export interface BaseTableConstructorOptions {
   modeParams?: any;
   canvasWidth?: number;
   canvasHeight?: number;
+  /**
+   * 'auto':和浏览器滚动行为一致 表格滚动到顶部/底部时 触发浏览器默认行为;
+   *  设置为 'none' 时, 表格滚动到顶部/底部时, 不再触发父容器滚动
+   * */
+  overscrollBehavior?: 'auto' | 'none';
 }
 export interface BaseTableAPI {
   /** 表格的行数 */
@@ -621,6 +633,12 @@ export interface BaseTableAPI {
   isFrozenRow: (col: number, row?: number) => boolean;
   isTopFrozenRow: (col: number, row?: number) => boolean;
   isBottomFrozenRow: (col: number, row?: number) => boolean;
+  /** 获取表格body部分的显示单元格范围 */
+  getBodyVisibleCellRange: () => { rowStart: number; colStart: number; rowEnd: number; colEnd: number };
+  /** 获取表格body部分的显示列号范围 */
+  getBodyVisibleColRange: () => { colStart: number; colEnd: number };
+  /** 获取表格body部分的显示行号范围 */
+  getBodyVisibleRowRange: () => { rowStart: number; rowEnd: number };
 }
 export interface ListTableProtected extends IBaseTableProtected {
   /** 表格数据 */